+++ rl_zoo3/import_envs.py
@@ -1,5 +1,7 @@
-import gym
-from gym.envs.registration import register
+from typing import Optional
+
+import gymnasium as gym
+from gymnasium.envs.registration import register
 
 from rl_zoo3.wrappers import MaskVelocityWrapper
 
@@ -49,15 +51,15 @@ except ImportError:
     rocket_lander_gym = None
 
 try:
-    import gym_minigrid  # pytype: disable=import-error
+    import minigrid  # pytype: disable=import-error
 except ImportError:
-    gym_minigrid = None
+    minigrid = None
 
 
 # Register no vel envs
 def create_no_vel_env(env_id: str):
-    def make_env():
-        env = gym.make(env_id)
+    def make_env(render_mode: Optional[str] = None):
+        env = gym.make(env_id, render_mode=render_mode)
         env = MaskVelocityWrapper(env)
         return env