+++ rl_zoo3/record_video.py
@@ -90,6 +90,9 @@ if __name__ == "__main__":
     if args.env_kwargs is not None:
         env_kwargs.update(args.env_kwargs)
 
+    # Force rgb_array rendering (gym 0.26+)
+    env_kwargs.update(render_mode="rgb_array")
+
     env = create_test_env(
         env_name.gym_id,
         n_envs=n_envs,
@@ -154,10 +157,10 @@ if __name__ == "__main__":
                 episode_start=episode_starts,
                 deterministic=deterministic,
             )
-            obs, _, dones, _ = env.step(action)  # type: ignore[assignment]
-            episode_starts = dones
             if not args.no_render:
                 env.render()
+            obs, _, dones, _ = env.step(action)  # type: ignore[assignment]
+            episode_starts = dones
     except KeyboardInterrupt:
         pass