+++ rl_zoo3/train.py
@@ -5,7 +5,8 @@ import os
 import time
 import uuid
 
-import gym
+import gym as gym26
+import gymnasium as gym
 import numpy as np
 import stable_baselines3 as sb3
 import torch as th
@@ -157,7 +158,9 @@ def train() -> None:
         importlib.import_module(env_module)
 
     env_id = args.env
-    registered_envs = set(gym.envs.registry.env_specs.keys())  # pytype: disable=module-attr
+    registered_envs = set(gym.envs.registry.keys())  # pytype: disable=module-attr
+    # Add gym 0.26 envs
+    registered_envs.update(gym26.envs.registry.keys())  # pytype: disable=module-attr
 
     # If the environment is not found, suggest the closest match
     if env_id not in registered_envs: